MAINT: use agm from xsf and remove cython
