+++ scipy/special/_generate_pyx.py
@@ -81,7 +81,7 @@ special_ufuncs = [
     '_normalized_gen_harmonic', '_scaled_exp1', '_sinpi', '_von_mises_cdf',
     '_spherical_jn', '_spherical_jn_d', '_spherical_yn', '_spherical_yn_d',
     '_spherical_in', '_spherical_in_d', '_spherical_kn', '_spherical_kn_d',
-    'airy', 'airye', 'bei', 'beip', 'ber', 'berp', 'binom',
+    'agm', 'airy', 'airye', 'bei', 'beip', 'ber', 'berp', 'binom',
     'exp1', 'expi', 'expit', 'exprel', 'gamma', 'gammaln', 'hankel1',
     'hankel1e', 'hankel2', 'hankel2e', 'hyp2f1', 'it2i0k0', 'it2j0y0', 'it2struve0',
     'itairy', 'iti0k0', 'itj0y0', 'itmodstruve0', 'itstruve0',
